Encapsulating the 'computer' in a separate class.
