+++ .pryrc
@@ -1,1 +1,2 @@
 require './lib/game.rb'
+require './lib/computer.rb'
