+++ lib/computer.rb
@@ -1,0 +1,5 @@
+class Computer
+  def weapon
+    [:rock, :paper, :scissors, :spock, :lizard].sample
+  end
+end
