+++ spec/coumputer_spec.rb
@@ -1,0 +1,10 @@
+require 'computer'
+
+describe Computer do
+  subject(:computer) { described_class.new }
+  let(:options) { [:rock, :paper, :scissors, :spock, :lizard] }
+
+  it 'returns random variable within class' do
+    expect(options).to include(computer.weapon)
+  end
+end
